feat: rewrite sd-notification documentation
